fixed comment; moved pileup.element to separate class PileupElement
