+++ guacamole-core/src/main/scala/org/bdgenomics/guacamole/pileup/Pileup.scala
@@ -19,23 +19,20 @@
 package org.bdgenomics.guacamole.pileup
 
 import org.bdgenomics.adam.rich.DecadentRead
-import scala.collection.JavaConversions._
-import net.sf.samtools.{ CigarElement, CigarOperator, TextCigarCodec }
-import scala.annotation.tailrec
-import org.bdgenomics.guacamole.{ CigarUtils, Common }
+import org.bdgenomics.guacamole.Common
 
 /**
- * A [[Pileup]] at a locus contains a sequence of [[Pileup.Element]] instances, one for every read that overlaps that
- * locus. Each [[Pileup.Element]] specifies the base read at the given locus in a particular read. It also keeps track
+ * A [[Pileup]] at a locus contains a sequence of [[PileupElement]] instances, one for every read that overlaps that
+ * locus. Each [[PileupElement]] specifies the base read at the given locus in a particular read. It also keeps track
  * of the read itself and the offset of the base in the read.
  *
  *
  * @param locus The locus on the reference genome
  *
- * @param elements Sequence of [[Pileup.Element]] instances giving the sequenced bases that align to a particular
+ * @param elements Sequence of [[PileupElement]] instances giving the sequenced bases that align to a particular
  *                 reference locus, in arbitrary order.
  */
-case class Pileup(locus: Long, elements: Seq[Pileup.Element]) {
+case class Pileup(locus: Long, elements: Seq[PileupElement]) {
   /** The first element in the pileup. */
   lazy val head = {
     assume(!elements.isEmpty, "Empty pileup")
@@ -79,7 +76,7 @@ case class Pileup(locus: Long, elements: Seq[Pileup.Element]) {
       "New locus (%d) must be greater than current locus (%d)".format(newLocus, locus))
     val reusableElements = elements.filter(element => Common.overlapsLocus(element.read.record, newLocus))
     val updatedElements = reusableElements.map(_.elementAtGreaterLocus(newLocus))
-    val newElements = newReads.map(Pileup.Element(_, newLocus))
+    val newElements = newReads.map(PileupElement(_, newLocus))
     Pileup(newLocus, updatedElements ++ newElements)
   }
 
@@ -95,7 +92,7 @@ object Pileup {
    */
   def pileupsAtLoci(locusAndReads: Iterator[(Long, Iterable[DecadentRead])]): Iterator[Pileup] = {
 
-    val empty: Seq[Pileup.Element] = Seq()
+    val empty: Seq[PileupElement] = Seq()
     val initialEmpty = Pileup(0, empty)
 
     val iterator = locusAndReads.scanLeft(initialEmpty)((prevPileup: Pileup, pair) => {
@@ -114,201 +111,10 @@ object Pileup {
    * @return A [[Pileup]] at the given locus.
    */
   def apply(reads: Seq[DecadentRead], locus: Long): Pileup = {
-    val elements = reads.filter(read => Common.overlapsLocus(read.record, locus)).map(Element(_, locus))
+    val elements = reads.filter(read => Common.overlapsLocus(read.record, locus)).map(PileupElement(_, locus))
     val pileup = Pileup(locus, elements)
     assert(pileup.locus == locus, "New pileup has locus %d but exepcted %d".format(pileup.locus, locus))
     pileup
-  }
-
-  /**
-   * A [[Pileup.Element]] represents a particular nucleotide sequenced by a particular read at a particular
-   * reference locus.
-   *
-   * @param read The read this [[Element]] is coming from.
-   * @param locus The reference locus.
-   * @param readPosition The offset into the sequence of bases in the read that this element corresponds to.
-   * @param remainingReadCigar A list of remaining parsed cigar elements of this read.
-   * @param cigarElementLocus The reference START position of the cigar element.
-   *                          If the element is an INSERTION this the PRECEDING reference base
-   * @param indexInCigarElements Which cigar element in the read this element belongs to.
-   * @param indexWithinCigarElement The offset of this element within the current cigar element
-   */
-  case class Element(
-      read: DecadentRead,
-      locus: Long,
-      readPosition: Long,
-      remainingReadCigar: List[CigarElement],
-      cigarElementLocus: Long,
-      indexInCigarElements: Long,
-      indexWithinCigarElement: Long) {
-
-    assume(locus >= read.record.getStart)
-    assume(locus < read.record.end.get)
-    assume(read.record.mdTag.isDefined, "Record has no MDTag.")
-
-    lazy val cigarElement = remainingReadCigar.head
-    lazy val nextCigarElement = if (remainingReadCigar.tail.isEmpty) None else Some(remainingReadCigar.tail.head)
-
-    /*
-     * True if last base of the current cigar element
-     */
-    def isFinalCigarBase: Boolean = indexWithinCigarElement == cigarElement.getLength - 1
-
-    lazy val alignment: Alignment = {
-      val cigarOperator = cigarElement.getOperator
-      val nextBaseCigarElement = if (isFinalCigarBase) nextCigarElement else Some(cigarElement)
-      val nextBaseCigarOperator = nextBaseCigarElement.map(_.getOperator)
-      (cigarOperator, nextBaseCigarOperator) match {
-        //To process insertions we connect the final base of a match to the rest of the insertion
-        case (CigarOperator.M, Some(CigarOperator.I)) | (CigarOperator.EQ, Some(CigarOperator.I)) | (CigarOperator.I, _) =>
-          val startReadOffset: Int = readPosition.toInt
-          val endReadOffset: Int = readPosition.toInt + CigarUtils.getReadLength(nextBaseCigarElement.get) + 1
-          val bases = read.record.getSequence.subSequence(startReadOffset, endReadOffset).toString
-          val qualities = read.record.qualityScores.slice(startReadOffset, endReadOffset)
-          Insertion(bases, qualities)
-        case (CigarOperator.M, _) | (CigarOperator.EQ, _) | (CigarOperator.X, _) =>
-          val base: Char = read.record.getSequence.charAt(readPosition.toInt)
-          val quality = read.record.qualityScores(readPosition.toInt)
-          if (read.record.mdTag.get.isMatch(locus)) {
-            Match(base, quality)
-          } else {
-            Mismatch(base, quality)
-          }
-        case (CigarOperator.D, _) | (CigarOperator.S, _) | (CigarOperator.N, _) | (CigarOperator.H, _) => Deletion()
-        case (CigarOperator.P, _) =>
-          throw new AssertionError("`P` CIGAR-ops should have been ignored earlier in `findNextCigarElement`")
-      }
-    }
-
-    /* If you only care about what the CigarOperator was at this position, but not its
-     * associated sequence, then you can just one of these state variables.
-     */
-    lazy val isInsertion = alignment match { case Insertion(_, _) => true; case _ => false }
-    lazy val isDeletion = alignment match { case Deletion() => true; case _ => false }
-    lazy val isMismatch = alignment match { case Mismatch(_, _) => true; case _ => false }
-    lazy val isMatch = alignment match { case Match(_, _) => true; case _ => false }
-
-    /**
-     * The sequenced nucleotides at this element.
-     *
-     * If the current element is a deletion, then return the empty string. If it's
-     * an insertion, then this will be a string of length >= 1: the contents of
-     * the inserted sequence starting at the current locus. Otherwise, this is
-     * a string of length 1.
-     */
-    lazy val sequenceRead: String = alignment match {
-      case Deletion()          => ""
-      case Match(base, _)      => base.toString
-      case Mismatch(base, _)   => base.toString
-      case Insertion(bases, _) => bases
-    }
-
-    lazy val singleBaseRead: Char = alignment match {
-      case Match(base, _)                           => base
-      case Mismatch(base, _)                        => base
-      case Insertion(bases, _) if bases.length == 1 => bases.charAt(0)
-      case other =>
-        throw new AssertionError("Not a match, mismatch, or single nucleotide insertion: " + other.toString)
-    }
-
-    /*
-     * Phred-scaled base quality score
-     * For indels this uses a specific heuristic
-     * Deletion quality is equivalent to the mapping quality
-     * Insertion quality is equivalent to worse base quality in the read (alternatively override and use average)
-     */
-    lazy val qualityScore: Int = alignment match {
-      case Deletion()              => read.record.getMapq //Heuristic for deletion quality,
-      // in the lack of other evidence it is the mapping quality
-      case Match(_, quality)       => quality
-      case Mismatch(_, quality)    => quality
-      case Insertion(_, qualities) => qualities.min // Heuristic for insertion quality,
-      // In the lack of other evidence, it is the quality of the worst base
-    }
-
-    /**
-     * Returns a new [[Pileup.Element]] of the same read at a different locus.
-     *
-     * To enable an efficient implementation, newLocus must be greater than the current locus.
-     *
-     * @param newLocus The desired locus of the new [[Pileup.Element]]. It must be greater than the current locus, and
-     *                 not past the end of the current read.
-     *
-     * @return A new [[Pileup.Element]] at the given locus.
-     */
-    def elementAtGreaterLocus(newLocus: Long): Element = {
-      if (newLocus == locus) { return this }
-
-      assume(newLocus > locus, "Can't rewind to locus %d from %d. Pileups only advance.".format(newLocus, locus))
-      val readEndPos = read.record.end.get
-      assume(newLocus < readEndPos, "This read stops at position %d. Can't advance to %d".format(readEndPos, newLocus))
-
-      val currentCigarReadPosition = if (cigarElement.getOperator.consumesReadBases()) {
-        readPosition - indexWithinCigarElement
-      } else {
-        readPosition
-      }
-      // Iterate through the remaining cigar elements to find one overlapping the current position
-      @tailrec
-      def getCurrentElement(remainingCigarElements: List[CigarElement], cigarReadPosition: Long, cigarReferencePosition: Long, cigarElementIndex: Long): Element = {
-        if (remainingCigarElements.isEmpty) {
-          throw new RuntimeException(
-            "Couldn't find cigar element for locus %d, cigar string only extends to %d".format(newLocus, cigarReferencePosition))
-        }
-        val nextCigarElement = remainingCigarElements.head
-        val cigarOperator = nextCigarElement.getOperator
-        val cigarElementReadLength = CigarUtils.getReadLength(nextCigarElement)
-        val cigarElementReferenceLength = CigarUtils.getReferenceLength(nextCigarElement)
-        // The 'P' (padding) operator is used to indicate a deletion-in-an-insertion. This only comes up when the
-        // aligner attempted not only to align reads to the reference, but also to align inserted sequences within reads
-        // to each other. In particular, a de novo assembler would automatically be doing this.
-        // We ignore this operator, since our simple Alignment handling code does not expose within-insertion alignments.
-        // See: http://davetang.org/wiki/tiki-index.php?page=SAM
-        if (cigarOperator != CigarOperator.P) {
-          val currentElementEnd = cigarReferencePosition + cigarElementReferenceLength
-          if (currentElementEnd > newLocus) {
-            val offset = newLocus - cigarReferencePosition
-            val finalReadPos = if (cigarOperator.consumesReadBases) cigarReadPosition + offset else cigarReadPosition
-            return Element(
-              read,
-              newLocus,
-              finalReadPos,
-              remainingCigarElements,
-              cigarReferencePosition,
-              cigarElementIndex,
-              offset)
-          }
-        }
-        getCurrentElement(remainingCigarElements.tail, cigarReadPosition + cigarElementReadLength, cigarReferencePosition + cigarElementReferenceLength, cigarElementIndex + 1)
-      }
-
-      getCurrentElement(remainingReadCigar, currentCigarReadPosition, cigarElementLocus, indexInCigarElements)
-    }
-  }
-
-  object Element {
-    /**
-     * Create a new [[Pileup.Element]] backed by the given read at the specified locus. The read must overlap the locus.
-     *
-     */
-    def apply(read: DecadentRead, locus: Long): Element = {
-      assume(read.isAligned)
-      assume(locus >= read.record.start)
-      assume(read.record.end.isDefined)
-      assume(locus < read.record.end.get)
-
-      val cigar = TextCigarCodec.getSingleton.decode(read.record.getCigar.toString).getCigarElements.toList
-      val startElement = Element(
-        read = read,
-        locus = read.record.getStart,
-        readPosition = 0,
-        remainingReadCigar = cigar,
-        cigarElementLocus = read.record.getStart,
-        indexInCigarElements = 0,
-        indexWithinCigarElement = 0)
-
-      startElement.elementAtGreaterLocus(locus)
-    }
   }
 }
